==> 2018/2018-06-28-数据分析流程/ref/案例研究2_汽车燃油性能对比_2008vs2018/11-fix-datatypes-mpg-greenhouse.ipynb <==
# 修正数据类型第 3 部分

在最后一部分，你将修正 mpg 和温室气体得分列的数据类型。

完成这些最终修正后，请检查两个数据集中所有特征的数据类型，以确认我们之前指定的所有更改成功实施。这里再次提供它们以供参考:
修正 cyl 数据类型

    2008: 从字符串中提取整型。
    2018: 将浮点型转换为整型。

修正 air_pollution_score 数据类型

    2008: 将字符串转换为浮点型。
    2018: 将整型转换为浮点型。

修正 city_mpg、hwy_mpg、cmb_mpg 数据类型

    2008 and 2018: 将字符串转换为浮点型。

修正 greenhouse_gas_score 数据类型

    2008: 将浮点型转换为整型。



# load datasets
import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

## Fix `city_mpg`, `hwy_mpg`, `cmb_mpg` datatypes
    2008 and 2018: convert string to float

# convert mpg columns to floats
mpg_columns = ['city_mpg', 'hwy_mpg', 'cmb_mpg']
for c in mpg_columns:
    df_18[c] = df_18[c].astype(float)
    df_08[c] = df_08[c].astype(float)

## Fix `greenhouse_gas_score` datatype
    2008: convert from float to int

# convert from float to int
df_08['greenhouse_gas_score'] = df_08['greenhouse_gas_score'].astype(int)

## All the dataypes are now fixed! Take one last check to confirm all the changes.

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# Save your new CLEAN datasets as new files!
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
